fix(engine) #5637: address the code review on #5644
- Document why SCHEMA_CHANGE reports the whole current picture rather than the delta caused
  by the new index. Attributing each finding to an index would mean matching on message text
  or teaching checkSuitability() to attribute, and would report a blocked state as partly
  acceptable; the line count is bounded by the indexes on the type and CREATE INDEX is rare.

- Document why the new saveConfiguration() is deliberately NOT guarded on "the strategy
  actually changed", unlike several sibling mutators. saveConfiguration() reports an
  IOException by logging SEVERE and returning, so memory and schema.json can diverge, and
  re-issuing the same ALTER TYPE is the operator's repair for exactly that. A skip-if-unchanged
  guard reads the in-memory value and turns the repair into a silent no-op.

- Extract the log-capture helper this PR had duplicated across the two partitioning test
  classes into WarningCapture, and record there that the logger swap is process-global and so
  assumes the sequential execution the engine suite uses.

- Drop an HTML <p> that had been copied into a non-Javadoc // comment in TypeIndexBuilder.
